requirements.txt
sqlalchemy==1.4.22
apscheduler==3.7.0
scrapy==2.4.1
itemadapter==0.3.0
python-dotenv==0.19.2